Moved init of Config class away from ctor
Now the initialization must be done by the caller. This is necessary because Python doesn't have function/method overloading and two different ctors would be necessary to initialize the Config - the first ctor would have to parse the metadata, command and schedule from a string. The second one would accept explicit parameters.